change data type to FP16 into model

+++ jetson.py
@@ -224,7 +224,7 @@ class ObjectDetection():
 		)
 
 		predictions = self.model(
-			processed_frame.cpu().numpy()
+			processed_frame.cpu().numpy().astype(numpy.float16)
 		)
 
 		predictions = postprocess(
